return error when none of the given config sources exist, improves #3
When no single configuration source exists we can't build a map of
repositories which is used later on to translate how USE references
can be accessed. This would result in a "NoRepository" error which
would give the user a wrong hint instead of pointing out that we
can't read any of the config files.