Make default configuration builtin instead of shipping it as a file
Refactor configuration API while doing so. There are now two ways to
create Configuration instances:

* config.Default() -> The builtin supporting BitBucket and GitHub.
* config.Empty() -> An empty configuration with no repositories declared.

The configuration instances retrieved by these functions can then be
merged with files by calling Merge().

If no `-config <file>` argument is passed, the default config is used
and merged with files in the searchpath as defined in #3. If a file
is explicitely specified, it is used as single authorative source, the
search path is not checked.

See discussion in issue #5