implement clean subcommand, fixes #9
This introduces a new subcommand to clean up and remove the vendor
directory. Also, it renames the used vendor directory to doget_modules.

Additionally, a switch -clean for the transform subcommand is
introduced which allows to remove the vendor directory immediatly
after the transformation is done, leaving no traces that the command
has run except the newly created Dockerfile.