Implement PROVIDES and USE...AS instructions as defined in #12
The `PROVIDES` instruction declares a trait provides a certain image
when included, e.g. "php:latest". Other traits can then happily extend
from these in their `FROM` statement.

The `AS` extension to the `USE` instruction is a shorthand for when
we do not have control over the trait we're including. Adding it
behaves the same as if the trait had included a `PROVIDES` instruction
with the same image.