inject actual commands
By injecting actual commands we can have a test which asserts and
describes the behaviour of the build command.

Currently the docker client is injected. However, it might be better
to inject a function which creates the docker client, so it becomes
possible to define the docker binary (or docker http daemon) as
command line argument to the build command.

At the moment this introduces a separate docker package. It should
be discussed on whether to keep this, or if this should be part of
the command/build package itself.